refactor(core): extract daemon lease staleness/freshness logic into daemon_lease
daemon.rs is a 3372-line god-file, and the recurring daemon-recovery bugs
(#9128 lease-bound stops, #8967 fresh-daemon reconnect, #8897 abandoned
reservations) keep landing in its lease staleness-decision logic. This isolates
that cluster into a new daemon/daemon_lease submodule so the invariants are
reviewable in one place, separate from the HTTP-serving spine.

Moves the pure (no JobStore / no serving) decision logic:
- DaemonLeaseValidation (the typed verdict the serving code consumes)
- validate_lease_file (the entry point), stale_lease, daemon_state_identity,
  freshness_report_from_validation, runtime_snapshot_stale_reason

Pure move, no behavior change. DaemonLeaseValidation + its fields become
pub(crate) (the 6 serving-side call sites read every field); the shared deps it
needs (DAEMON_LEASE_SCHEMA, runtime_path_fingerprint, current_binary_sha256,
read_termination_evidence) are made pub(crate) and referenced via super::. The
serving spine imports validate_lease_file/freshness_report_from_validation/
daemon_state_identity back. daemon.rs -214 lines (3372 -> 3164).

Verified: homeboy-core --lib + --tests clean, no unused-import warnings. The 8
daemon::tests (which exercise validate_lease_file/freshness) pass. The 4
real-OS-process/broker batch failures (daemon_operation_lock_recovers...,
dead_daemon_recovery..., remote_runner_broker_redacts...) are pre-existing
#8964-class flakes and fail identically on origin/main without this change.
